requirements.txt
discord.py @ git+https://github.com/Rapptz/discord.py
python-dotenv
mysql-connector-python
